requirements.txt
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
openai==1.66.3
